Make sure new cluster's max compaction lag is of 14 days
As part of the global project for customer data deletion, log
compaction has a new maximum lag of 14 days, after which compaction
happens.

This change has already been applied dynamically to our production and
test cluster, but in the case a new cluster gets deployed in the
future, in order to avoid a "regression" from sneaking in, we better
build bubuku with this property by default, that way our current
solution becomes future proof.