Disable unclean leader election
Unclean leader election used to be enabled by default until Kafka 0.11.0.0.

But even before that Nakadi explicitly configured unclean leader
election as enabled for a number of historical reasons. One Nakadi was
first released in 2016, Kafka used to be deployed using stups
autoscaling groups and the volume storing event data used to be
considered ephemeral, e.g. we considered that failures in the ec2
instance could lead to permanent loss of the accompaning data volume.

Such a MTTR was unacceptable and we made the decision to turn on
unclean leader election.

In such a scenario, losing 2 kafka brokers could result in a mean time
to recover of more than a day.

This all changed once we started to treat EBS volumes as something we
can "reattach", meaning, the chances of losing such a volume decreased
dramatically as failures on ec2 instances are much more frequent than
failures resulting in permanent loss of an EBS volume. This change was
introduced somewhere in 2017 and is documented in this blog post https://engineering.zalando.com/posts/2017/10/reattaching-kafka-ebs-in-aws.html

In this new world were MTTR would be much faster, it makes sense to
have unclean.leader.election turned off, since most users expect
durability and consistency over availability.

That all said, losing 2 replicas at the same time is a very rare event
that probably happened only a couple of times during Nakadi's almost 5
year long history during severe outage of AWS availability zones
before black friday 2019 and during a more recenve incident in
2020. Though we cannot confirm as there are no traces to be
investigated.

A few notable mentions:

> Notable changes in 0.11.0.0 Unclean leader election is now disabled
> by default. The new default favors durability over
> availability. Users who wish to to retain the previous behavior
> should set the broker config unclean.leader.election.enable to true.

> Notable changes in 2.1.0
> Unclean leader election is automatically
> enabled by the controller when unclean.leader.election.enable config
> is dynamically updated by using per-topic config override.

This means that in between 0.11.0.0 and 2.1.0 in order to force a
kafka broker to enable unclean leader election on a per topic basis in
order to recover from a catastrophic scenario would require forcing
the election of kafka controller. This was not documented in Kafka and
they probably made significant money with consulting clients on how to
recover from catastrophic situations before they finally fix it and
mentioned it casually in their release docs. That goes to show that
our decision of keeping unclean leader election disabled might have
spared us of a couple of terrible incidents where Nakadi could stay
down for hours, which would have harmed our reputation to offer a high
level service and possibly we wouldn't even exist anymore, as we
wouldn't be considered reliable enough for mission critical usage.