Reduce leader imbalance
When performing a rolling restart, one of the conditions for bubuku to
restart a broker is that the cluster is in a green state. This means
no under replicated partitions and no preferred replica imbalance.

Kafka deployment by default already enables automatic check for
imbalance and in our case this is configured to happen every 100
seconds. But by default it triggers only when 10% of partitions are
imbalanced for a broker (that's my understanding).

From Kafka docs:

```
leader.imbalance.per.broker.percentage
The ratio of leader imbalance allowed per broker. The controller would trigger a leader balance if it goes above this value per broker. The value is specified in percentage.

Type:	int
Default:	10
Valid Values:
Importance:	high
Update Mode:	read-only
```

This change is an atempt to cause it to trigger more often in order to
avoid the rolling restart from getting stuck and requireing manual fix
before it can continue.