Change default cooldown to 10 minutes
We learned from experience that rolling restarts are less disruptive
when performed slowly over time. This gives consumers enough time to
breath (rebuild state or properly fetch up-to-date metadata and
dispatch potentially accumulated backlog) in between restarts.

We are making the default 10 minutes so that the next person to
perform a rolling restart doesn't have to find out the hard way that
speedy rolling restarts may cause incidents.